cleanup

+++ hive/indexer/cache.py
@@ -255,13 +255,16 @@ def rebuild_cache():
 
 # the feed cache allows for efficient querying of blogs+reblogs. this method
 # efficiently builds the feed cache after the initial sync.
-def rebuild_feed_cache():
-    print("Rebuilding hive_feed_cache")
-    query("TRUNCATE TABLE hive_feed_cache")
-    query("INSERT INTO hive_feed_cache SELECT author account, id post_id, created_at "
-        "FROM hive_posts WHERE depth = 0 AND is_deleted = 0")
+def rebuild_feed_cache(truncate = True):
+    print("*** Rebuilding hive_feed_cache ***")
+    if truncate:
+        query("TRUNCATE TABLE hive_feed_cache")
+
     query("INSERT IGNORE INTO hive_feed_cache "
-        "SELECT account, post_id, created_at FROM hive_reblogs")
+            "SELECT author account, id post_id, created_at "
+            "FROM hive_posts WHERE depth = 0 AND is_deleted = 0")
+    query("INSERT IGNORE INTO hive_feed_cache "
+            "SELECT account, post_id, created_at FROM hive_reblogs")
 
 
 # identify and insert missing cache rows
